requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.9
aiofiles==23.2.1
slowapi==0.1.9
python-dotenv==1.0.1
pydantic==2.6.1
